requirements.txt
bs4
click
lxml
numpy
orhelper
pyyaml
